# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GlobalSearchParams": {
        "properties": {
          "min_cluster_size": {
            "default": 1,
            "title": "Min Cluster Size",
            "type": "integer"
          }
        },
        "title": "GlobalSearchParams",
        "type": "object"
      },
      "GlobalSearchRequest": {
        "additionalProperties": false,
        "properties": {
          "language": {
            "anyOf": [
              {
                "maxLength": 32,
                "minLength": 2,
                "pattern": "^[A-Za-z][A-Za-z \\-]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Answer language, e.g. 'russian'. Defaults to RAGU_API_LANGUAGE.",
            "title": "Language"
          },
          "params": {
            "$ref": "#/components/schemas/GlobalSearchParams",
            "description": "GlobalSearchEngine retrieval parameters: min_cluster_size"
          },
          "query": {
            "description": "Search query",
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "GlobalSearchRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Graph-RAG over one or more knowledge graphs.\n\nFour search modes — `global`, `local`, `naive`, `mix` — each in four shapes: an answer, retrieval without generation, a batch, and a Server-Sent Events stream. Graphs are addressed by name under `/v1/graphs/{graph_id}`; the flat `/v1/search/...` paths address the default graph and are kept for older clients.\n\nEvery response reports what actually ran (`engines`) and what it cost (`usage`). Errors share one envelope and carry `request_id`.",
    "title": "RAGU Service",
    "version": "0.0.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/search/global": {
      "post": {
        "deprecated": true,
        "operationId": "search_global_v1_search_global_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Capability unavailable for this graph"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "BUDGET_EXCEEDED: the request cannot fit its LLM budget. TOO_MANY_REQUESTS: every generation slot is taken"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Graph is not loaded / service not ready"
          }
        },
        "summary": "Search Global",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/global/batch": {
      "post": {
        "deprecated": true,
        "operationId": "batch_global_v1_search_global_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Capability unavailable for this graph"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "BUDGET_EXCEEDED: the request cannot fit its LLM budget. TOO_MANY_REQUESTS: every generation slot is taken"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Graph is not loaded / service not ready"
          }
        },
        "summary": "Batch Global",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/global/retrieve": {
      "post": {
        "deprecated": true,
        "operationId": "retrieve_global_v1_search_global_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalRetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Capability unavailable for this graph"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "BUDGET_EXCEEDED: the request cannot fit its LLM budget. TOO_MANY_REQUESTS: every generation slot is taken"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Graph is not loaded / service not ready"
          }
        },
        "summary": "Retrieve Global",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/global/stream": {
      "post": {
        "deprecated": true,
        "operationId": "stream_global_v1_search_global_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GlobalSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Capability unavailable for this graph"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "BUDGET_EXCEEDED: the request cannot fit its LLM budget. TOO_MANY_REQUESTS: every generation slot is taken"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Graph is not loaded / service not ready"
          }
        },
        "summary": "Stream Global",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/local": {
      "post": {
        "deprecated": true,
        "operationId": "search_local_v1_search_local_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Capability unavailable for this graph"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "BUDGET_EXCEEDED: the request cannot fit its LLM budget. TOO_MANY_REQUESTS: every generation slot is taken"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Graph is not loaded / service not ready"
          }
        },
        "summary": "Search Local",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/local/batch": {
      "post": {
        "deprecated": true,
        "operationId": "batch_local_v1_search_local_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Capability unavailable for this graph"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "BUDGET_EXCEEDED: the request cannot fit its LLM budget. TOO_MANY_REQUESTS: every generation slot is taken"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Graph is not loaded / service not ready"
          }
        },
        "summary": "Batch Local",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/local/retrieve": {
      "post": {
        "deprecated": true,
        "operationId": "retrieve_local_v1_search_local_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalRetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Capability unavailable for this graph"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "BUDGET_EXCEEDED: the request cannot fit its LLM budget. TOO_MANY_REQUESTS: every generation slot is taken"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Graph is not loaded / service not ready"
          }
        },
        "summary": "Retrieve Local",
        "tags": [
          "search"
        ]
      }
    },
...
  },
...
}